// ev-backoffice/client/src/app/account/profile/edit-preview-profile/models/degree.enum.ts
export enum DegreesValues {
  ASSOCIATE_OF_ARTS = 'ASSOCIATE_OF_ARTS',
  ASSOCIATE_OF_SCIENCE = 'ASSOCIATE_OF_SCIENCE',
  ASSOCIATE_OF_APPLIED_SCIENCE = 'ASSOCIATE_OF_APPLIED_SCIENCE',
  BACHELOR_OF_ARTS = 'BACHELOR_OF_ARTS',
  BACHELOR_OF_SCIENCE = 'BACHELOR_OF_SCIENCE',
  BACHELOR_OF_FINE_ARTS = 'BACHELOR_OF_FINE_ARTS',
  BACHELOR_OF_APPLIED_SCIENCE = 'BACHELOR_OF_APPLIED_SCIENCE',
  MASTER_OF_ARTS = 'MASTER_OF_ARTS',
  MASTER_OF_SCIENCE = 'MASTER_OF_SCIENCE',
  MASTER_OF_BUSINESS_ADMINISTRATION = 'MASTER_OF_BUSINESS_ADMINISTRATION',
  MASTER_OF_FINE_ARTS = 'MASTER_OF_FINE_ARTS',
  DOCTOR_OF_PHILOSOPHY = 'DOCTOR_OF_PHILOSOPHY',
  JURIS_DOCTOR = 'JURIS_DOCTOR',
  DOCTOR_OF_MEDICINE = 'DOCTOR_OF_MEDICINE',
  DOCTOR_OF_DENTAL_SURGERY = 'DOCTOR_OF_DENTAL_SURGERY',
}